src/turing_game_bot/requirements.txt
turing-bot-client
python-dotenv
openai
websockets
llama-cpp-python -C cmake.args="-DCMAKE_OSX_ARCHITECTURES=arm64 -DCMAKE_APPLE_SILICON_PROCESSOR=arm64 -DGGML_METAL=on"
huggingface-hub
ollama